Don't round range for DoubleSpinBoxDelegate to int
This fixes simple copy-paste error that the DoubleSpinBoxDelegate range
was stored as int, thus rounding min value from 0.2 to 0.

+++ qt-ui/modeldelegates.h
@@ -104,8 +104,8 @@ public:
 	DoubleSpinBoxDelegate(qreal min, qreal max, QObject *parent = 0);
 	virtual QWidget *createEditor(QWidget *parent, const QStyleOptionViewItem &option, const QModelIndex &index) const;
 private:
-	int min;
-	int max;
+	qreal min;
+	qreal max;
 };
 
 #endif // MODELDELEGATES_H
